requirements.txt
python-dotenv==1.0.1
strands-agents==1.7.1
streamlit==1.36.0
requests==2.32.3
pydantic==2.8.2
pytest==8.3.2
types-requests==2.32.0.20240712
mypy==1.11.1
